1. added validation 2. more tests 3. change aggregation value to be double instead of long

+++ src/main/java/com/mikerusoft/euroleague/model/Aggregation.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class Aggregation {
     private String command;
     private String field;
-    private long aggregatedValue;
+    private double aggregatedValue;
 }
